add a target to run the ipc tests if and when they all pass. Fix up a signed/unsigned error in the ipc_message unit test. Review URL: http://codereview.chromium.org/12946
git-svn-id: svn://svn.chromium.org/chrome/trunk/src@6358 0039d316-1c4b-4281-b951-d872f2087c98

+++ chrome/common/ipc_message_unittest.cc
@@ -17,7 +17,7 @@ TEST(IPCMessageTest, Serialize) {
     "#inva://idurl/",
   };
 
-  for (int i = 0; i < arraysize(serialize_cases); i++) {
+  for (size_t i = 0; i < arraysize(serialize_cases); i++) {
     GURL input(serialize_cases[i]);
     IPC::Message msg(1, 2, IPC::Message::PRIORITY_NORMAL);
     IPC::ParamTraits<GURL>::Write(&msg, input);
